Multiple receives per request Prefetch support; Calling thread only waits for non-prefetch content to arrive Memory Request Forwarding Multiple non-contiguous subscriptions per subtask

+++ source/code/src/pmCommand.cpp
@@ -372,7 +372,7 @@ pmCommunicatorCommand::remoteTaskAssignPacked::remoteTaskAssignPacked(pmLocalTas
 			lDeviceArray[i] = lDevices[i]->GetGlobalDeviceIndex();
 
 		devices.ptr = lDeviceArray;
-		devices.length = sizeof(uint) * taskStruct.assignedDeviceCount;
+		devices.length = (uint)(sizeof(uint)) * taskStruct.assignedDeviceCount;
 	}
 	else
 	{
